requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
tensorflow==2.15.0
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
Pillow==10.2.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6